feat: add new file ux.py containing UX helper modules
