+++ src/mmore/run_ragcli.py
@@ -16,16 +16,16 @@ logging.basicConfig(
 
 from mmore.profiler import enable_profiling_from_env, profile_function
 from mmore.rag.pipeline import RAGPipeline
-from mmore.ragcli_helper import (
+from mmore.ragcli_helper import TimingHandler
+from mmore.run_rag import RAGInferenceConfig
+from mmore.utils import load_config
+from mmore.ux import (
     Spinner,
-    TimingHandler,
     print_in_color,
     quiet_noisy_libs,
     str_green,
     str_in_color,
 )
-from mmore.run_rag import RAGInferenceConfig
-from mmore.utils import load_config
 
 
 class RagCLI:
